added operands and able to track 'the .json file from last successful compilation'

+++ entity/inc/label.h
@@ -7,13 +7,13 @@
 class Label : public Identifier
 {
 public:
-    Label(QString labelName, qint16 stmtLineNum);
+    Label(QString labelName, int stmtLineNum);
     ~Label();
     QJsonObject toJSON();
 
 private:
     QString value; // use string for now, change it later
-    qint16  stmtLineNum;
+    int  stmtLineNum;
 };
 
 #endif
